phan quyen cho phan SYS

+++ Sources/Source_Codes/FBDSource/FBD/Controllers/SYSReportingPeriodController.cs
@@ -25,6 +25,10 @@ namespace FBD.Controllers
         /// <returns>[Index] view</returns>
         public ActionResult Index()
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_VIEW, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             List<SystemReportingPeriods> lstPeriod = null;
             try
             {
@@ -51,6 +55,10 @@ namespace FBD.Controllers
         /// <returns>[Add] view</returns>
         public ActionResult Add()
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             return View();
         } 
 
@@ -71,6 +79,10 @@ namespace FBD.Controllers
         [HttpPost]
         public ActionResult Add(SystemReportingPeriods reportingPeriod)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 if (ModelState.IsValid)
@@ -124,6 +136,10 @@ namespace FBD.Controllers
         /// [Index] view: if ERROR</returns>
         public ActionResult Edit(string id)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             SystemReportingPeriods reportingPeriod = null;
 
             try
@@ -163,6 +179,10 @@ namespace FBD.Controllers
         [HttpPost]
         public ActionResult Edit(string id, SystemReportingPeriods reportingPeriod)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 if (ModelState.IsValid)
@@ -200,6 +220,10 @@ namespace FBD.Controllers
         /// <returns>[Index] view</returns>
         public ActionResult Delete(string id)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 int result = SystemReportingPeriods.DeleteReportingPeriod(id);
